Refactor chat store into focused persistence modules
